[java] Add command mapper that maps W3C WebDriver command to corresponding BiDi command
