+++ java/src/org/openqa/selenium/remote/BiDiCommandExecutor.java
@@ -17,179 +17,20 @@
 
 package org.openqa.selenium.remote;
 
-import static org.openqa.selenium.remote.DriverCommand.GET;
-import static org.openqa.selenium.remote.DriverCommand.PRINT_PAGE;
-
 import java.io.IOException;
-import java.io.StringReader;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.BiFunction;
-import org.openqa.selenium.UnsupportedCommandException;
-import org.openqa.selenium.bidi.Network;
-import org.openqa.selenium.bidi.Script;
-import org.openqa.selenium.bidi.browsingcontext.BrowsingContext;
-import org.openqa.selenium.bidi.browsingcontext.ReadinessState;
-import org.openqa.selenium.json.Json;
-import org.openqa.selenium.json.JsonInput;
-import org.openqa.selenium.json.TypeToken;
-import org.openqa.selenium.print.PageMargin;
-import org.openqa.selenium.print.PageSize;
-import org.openqa.selenium.print.PrintOptions;
 
 public class BiDiCommandExecutor implements CommandExecutor {
 
-  private static Json JSON = new Json();
-
   private final RemoteWebDriver driver;
-
-  private final AtomicReference<BrowsingContext> currentContext = new AtomicReference<>();
-  private final AtomicReference<Script> script = new AtomicReference<>();
-  private final AtomicReference<Network> network = new AtomicReference<>();
-
-  private final Map<String, BiFunction<Command, RemoteWebDriver, Response>> commandHandlerMap =
-      new HashMap<>();
-
-  // Each browsing context has an associated id, maintains state
-  // We will need to maintain a map of all the browsingContext to run further commands
-  // Switching between tabs etc might be tricky
-  private final Map<String, BrowsingContext> browsingContextMap = new ConcurrentHashMap<>();
+  private final BiDiCommandMapper commandMapper;
 
   public BiDiCommandExecutor(RemoteWebDriver driver) {
     this.driver = driver;
-    init(this.driver);
-  }
-
-  private void init(RemoteWebDriver driver) {
-    script.set(new Script(driver));
-    network.set(new Network(driver));
-
-    BrowsingContext parentContext = new BrowsingContext(driver, driver.getWindowHandle());
-    browsingContextMap.put(parentContext.getId(), parentContext);
-    currentContext.set(parentContext);
-
-    commandHandlerMap.put(
-        GET,
-        (command, webDriver) -> {
-          String pageLoadStrategy =
-              (String) webDriver.getCapabilities().getCapability("pageLoadStrategy");
-          currentContext
-              .get()
-              .navigate(
-                  (String) command.getParameters().get("url"),
-                  ReadinessState.getReadinessState(pageLoadStrategy));
-
-          // Required because W3C Classic sets the current browsing context to current top-level
-          // context on navigation
-          // This is crucial for tests that:
-          // Switch to frame -> find element -> navigate url -> find element (in BiDi it will try to
-          // find an element in the frame)
-          // But in WebDriver Classic it will try to find the element on the page navigated to
-          // So to avoid breaking changes, we need to add this step
-          // Refer: Pt 9 of https://www.w3.org/TR/webdriver2/#navigate-to
-          // Refer: https://w3c.github.io/webdriver-bidi/#command-browsingContext-navigate
-          driver.switchTo().window(currentContext.get().getId());
-          return new Response(webDriver.getSessionId());
-        });
-
-    commandHandlerMap.put(
-        PRINT_PAGE,
-        (command, webDriver) -> {
-          String result = currentContext.get().print(serializePrintOptions(command));
-          Response response = new Response(webDriver.getSessionId());
-          response.setValue(result);
-          return response;
-        });
+    commandMapper = new BiDiCommandMapper(driver);
   }
 
   @Override
   public Response execute(Command command) throws IOException {
-    // This is not optimal
-    // But each time Classic calls switchTo, we need to identify the current browsing context to run
-    // commands
-    // Maybe we need a way where every time switchTo commands are called, only then we update the
-    // current browsing context
-
-    String currentWindowHandle = driver.getWindowHandle();
-    BrowsingContext browsingContext =
-        browsingContextMap.computeIfAbsent(
-            currentWindowHandle, windowHandle -> new BrowsingContext(driver, windowHandle));
-
-    browsingContextMap.putIfAbsent(currentWindowHandle, browsingContext);
-    currentContext.set(browsingContext);
-    if (commandHandlerMap.containsKey(command.getName())) {
-      return commandHandlerMap.get(command.getName()).apply(command, driver);
-    } else {
-      throw new UnsupportedCommandException();
-    }
-  }
-
-  private PrintOptions serializePrintOptions(Command command) {
-    try (StringReader reader = new StringReader(JSON.toJson(command.getParameters()));
-        JsonInput input = JSON.newInput(reader)) {
-      PrintOptions printOptions = new PrintOptions();
-
-      input.beginObject();
-      while (input.hasNext()) {
-        switch (input.nextName()) {
-          case "page":
-            Map<String, Double> map = input.read(new TypeToken<Map<String, Double>>() {}.getType());
-            if (map.size() != 0) {
-              printOptions.setPageSize(new PageSize(map.get("height"), map.get("width")));
-            }
-            break;
-
-          case "orientation":
-            String orientation = input.read(String.class);
-            if (orientation.equals("portrait")) {
-              printOptions.setOrientation(PrintOptions.Orientation.PORTRAIT);
-            } else {
-              printOptions.setOrientation(PrintOptions.Orientation.LANDSCAPE);
-            }
-            break;
-
-          case "scale":
-            printOptions.setScale(input.read(Double.class));
-            break;
-
-          case "shrinkToFit":
-            printOptions.setShrinkToFit(input.read(Boolean.class));
-            break;
-
-          case "background":
-            printOptions.setBackground(input.read(Boolean.class));
-            break;
-
-          case "pageRanges":
-            List<String> pageRanges = input.read(new TypeToken<ArrayList<String>>() {}.getType());
-            printOptions.setPageRanges(pageRanges);
-            break;
-
-          case "margin":
-            Map<String, Double> marginMap =
-                input.read(new TypeToken<Map<String, Double>>() {}.getType());
-            if (marginMap.size() != 0) {
-              printOptions.setPageMargin(
-                  new PageMargin(
-                      marginMap.get("top"),
-                      marginMap.get("bottom"),
-                      marginMap.get("left"),
-                      marginMap.get("right")));
-            }
-            break;
-
-          default:
-            input.skipValue();
-            break;
-        }
-      }
-
-      input.endObject();
-      return printOptions;
-    }
+    return commandMapper.map(command).apply(command, driver);
   }
 }
